+++ src/main/java/net/meander/subtlyd/client/model/ModelProviderSD.java
@@ -77,7 +77,6 @@ public class ModelProviderSD extends FabricModelProvider {
         blockModelGenerator.createTrivialCube(BlocksSD.IRON_GRATE);
         blockModelGenerator.createDoublePlant(BlocksSD.REEDS, BlockModelGenerators.PlantType.NOT_TINTED);
         generateOverhangBlock(BlocksSD.WARPED_OVERHANG, blockModelGenerator);
-        blockModelGenerator.createDoor(BlocksSD.BASALT_DOOR);
     }
 
     @Override