+++ src/main/java/net/meander/subtlyd/mixin/client/entity/EntityMixin.java
@@ -1,6 +1,6 @@
 package net.meander.subtlyd.mixin.client.entity;
 
-import net.meander.subtlyd.util.CameraShake;
+import net.meander.subtlyd.camera.CameraShake;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.sounds.SoundEvent;
@@ -16,6 +16,6 @@ public class EntityMixin {
     @Inject(method = "playSound(Lnet/minecraft/sounds/SoundEvent;FF)V", at = @At("RETURN"))
     private void playSound(SoundEvent soundEvent, float f, float g, CallbackInfo ci) {
         Entity entity = (Entity) (Object) this;
-        CameraShake.shakeScreenFromSource(soundEvent, entity.blockPosition().getCenter(), 0);
+        CameraShake.shakeScreenFromSource(soundEvent, entity.blockPosition().getCenter());
     }
 }