+++ src/main/java/net/meander/subtlyd/mixin/common/server/ServerAdvancementManagerMixin.java
@@ -1,6 +1,6 @@
 package net.meander.subtlyd.mixin.common.server;
 
-import net.meander.subtlyd.util.data.AdvancementsSD;
+import net.meander.subtlyd.data.AdvancementsSD;
 import net.minecraft.advancements.Advancement;
 import net.minecraft.advancements.AdvancementRequirements;
 import net.minecraft.advancements.Criterion;